Refactor modules
- Split into env_parser, file_loaders, utils and exceptions modules
- Remove logtools module
- Add loaded values logging
